# pkcs5: optimize pbkdf2_hmac with larger passphrases and iteration counts