This significantly optimizes the mbedtls_pkcs5_pbkdf2_hmac function by
caching and reusing the initial HMAC digest state across all iterations.
This is especially evident when using passphrases longer than the digest
block size, which must first be digested for shortening before they can be
used a a key for the HMAC algorithm.

On an Intel Core i5-4690 CPU, using extreme parameters of SHA256 as the
digest, a passphrase length of 4096 bytes, a salt length of 128 bytes, an
iteration count of 1000000, and a derived key length of 128 bytes, the
runtime was reduced from 76 seconds to 3.75 seconds (a 95% reduction).  This
is comparable to OpenSSL's runtime of 2.75 seconds using the same
parameters.

The mbedtls_pkcs5_self_test() function, which uses SHA1 with much shorter
passphrases, salts, iterations, and keys, still had its runtime reduced from
10.1ms to 7.3ms (a 28% reduction).