@daverodgman 2022-02-11
This has been done in https://github.com/ARMmbed/mbed-crypto/pull/277.